Do not raise errors in BaseSession.wait_for_tasks for handled tasks
By default (and only way for now), a handled task is a task that
have a on_done callback defined. Such tasks do not report errors
when we wait for them from the session, as it is likely that we
used that callback to handle the error on our own.